uch shabbos

+++ game.h
@@ -12,7 +12,7 @@
 
 class Game{
 public:
-	Game(char*, int player_types[2], double t_limits[2], bool, bool, int hnum[2]);
+	Game(char*, int player_types[2], int t_limits[2], bool, bool, int hnum[2]);
 	void play(int);
 private:
 	bool prints;
